Added feature to delete all future occurences of recurring expense

+++ Kaching/Services/ExpenseService.cs
@@ -73,15 +73,22 @@ namespace Kaching.Services
 
         public async Task DeleteRecurringExpense(int expenseEventId)
         {
-            var expenseEvent = await _expenseRepository.GetExpenseById(expenseEventId);
-            var expense = await _baseExpenseRepository.GetBaseExpenseById(expenseEvent.ExpenseId);
+            var expense = await _expenseRepository.GetExpenseById(expenseEventId);
+            var baseExpense = await _baseExpenseRepository.GetBaseExpenseById(expense.BaseExpenseId);
 
-            foreach (var item in expense.Expenses)
+            foreach (var item in baseExpense.Expenses)
             {
-                _expenseRepository.DeleteExpense(item);
+                if (item.PaymentDate > expense.PaymentDate )
+                {
+                    _expenseRepository.DeleteExpense(item);
+                }
             }
 
-            _baseExpenseRepository.DeleteBaseExpense(expense);
+            // Deletes base expense class when there are no expenses belonging to it
+            if (expense.BaseExpense.Expenses.Count <= 1)
+            {
+                _baseExpenseRepository.DeleteBaseExpense(baseExpense);
+            }
             await _baseExpenseRepository.SaveAsync();
         }
 
@@ -89,7 +96,7 @@ namespace Kaching.Services
         {
             var expenseEvent = await _expenseRepository.GetExpenseById(expenseEventId);
 
-            // Deletes base expense class when there are no 
+            // Deletes base expense class when there are no expenses belonging to it
             if (expenseEvent.BaseExpense.Expenses.Count <= 1)
             {
                 var baseExpense = expenseEvent.BaseExpense;
